Extract signal handling into a dedicated source file
